% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_riclpm.R
\name{latent_riclpm}
\alias{latent_riclpm}
\title{Function to generate syntax for multiple indicator RICLPM
Inputs:
v = name want to give each variable. E.g., c("Var1","Var2", "Var3")
items = item range for each variable. E.g., c(3, 4, 6)
time = numeric number of time waves E.g., 4}
\usage{
latent_riclpm(v, items, time, constraint, resid)
}
\description{
Function to generate syntax for multiple indicator RICLPM
Inputs:
v = name want to give each variable. E.g., c("Var1","Var2", "Var3")
items = item range for each variable. E.g., c(3, 4, 6)
time = numeric number of time waves E.g., 4
}
